Snake casing and renaming encoders per code standards
