+++ src/nupic/encoders/adaptive_scalar.capnp
