+++ src/nupic/encoders/adaptive_scalar.py
