+++ src/nupic/encoders/logarithm.capnp
