+++ src/nupic/encoders/logarithm.py
