+++ src/nupic/encoders/pass_through.py
