+++ src/nupic/encoders/sdr_category.capnp
