+++ src/nupic/encoders/sdr_category.py
