+++ src/nupic/encoders/sparse_pass_through.capnp
